# swagger.json
{
	"basePath": "/",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"ILakeModel": {
			"properties": {
				"id": {
					"type": "number",
					"format": "double",
					"x-nullable": true
				},
				"name": {
					"type": "string"
				},
				"salmon": {
					"type": "number",
					"format": "double"
				},
				"tuna": {
					"type": "number",
					"format": "double"
				}
			},
			"required": [
				"name",
				"salmon",
				"tuna"
			],
			"type": "object"
		}
	},
	"info": {
		"title": "lakedb-project",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		}
	},
	"paths": {
		"/lakes": {
			"get": {
				"operationId": "GetAll",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/ILakeModel"
							}
						}
					}
				},
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "requestBody",
						"required": true,
						"schema": {
							"$ref": "#/definitions/ILakeModel"
						}
					}
				]
			}
		},
		"/lakes/{id}": {
			"get": {
				"operationId": "Get",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/ILakeModel"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			},
			"put": {
				"operationId": "Update",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "requestBody",
						"required": true,
						"schema": {
							"$ref": "#/definitions/ILakeModel"
						}
					}
				]
			},
			"delete": {
				"operationId": "Remove",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			}
		}
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {}
}